Rejigger _pty_size to follow recent error() changes
