+++ fabric/operations.py
@@ -21,40 +21,13 @@ from fabric.network import needs_host, ssh
 from fabric.sftp import SFTP
 from fabric.state import env, connections, output, win32, default_channel
 from fabric.thread_handling import ThreadHandler
-from fabric.utils import abort, indent, warn, puts, handle_prompt_abort, error
+from fabric.utils import abort, indent, warn, puts, handle_prompt_abort, error, _pty_size
 
 # For terminal size logic below
 if not win32:
     import fcntl
     import termios
     import struct
-
-
-def _pty_size():
-    """
-    Obtain (rows, cols) tuple for sizing a pty on the remote end.
-
-    Defaults to 80x24 (which is also the 'ssh' lib's default) but will detect
-    local (stdout-based) terminal window size on non-Windows platforms.
-    """
-    rows, cols = 24, 80
-    if not win32 and sys.stdout.isatty():
-        # We want two short unsigned integers (rows, cols)
-        fmt = 'HH'
-        # Create an empty (zeroed) buffer for ioctl to map onto. Yay for C!
-        buffer = struct.pack(fmt, 0, 0)
-        # Call TIOCGWINSZ to get window size of stdout, returns our filled
-        # buffer
-        try:
-            result = fcntl.ioctl(sys.stdout.fileno(), termios.TIOCGWINSZ,
-                buffer)
-            # Unpack buffer back into Python data types
-            rows, cols = struct.unpack(fmt, result)
-        # Deal with e.g. sys.stdout being monkeypatched, such as in testing.
-        # Or termios not having a TIOCGWINSZ.
-        except AttributeError:
-            pass
-    return rows, cols
 
 
 def _shell_escape(string):
